feat: init pwa
