+++ apps/mobile/index.html
@@ -1,10 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="/im-chat.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>IM-Chat</title>
+    <meta name="description" content="a chat tool">
+    <link rel="icon" type="image/png" href="/icon.svg" />
+    <link rel="apple-touch-icon" href="/icon_180.png" sizes="180x180">
+    <meta name="theme-color" content="#ffffff">
+    <meta charset="UTF-8" />
   </head>
   <body>
     <div id="app"></div>
