+++ apps/mobile/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@vueuse/core": "^9.13.0",
     "axios": "^1.3.4",
+    "chat-logic": "workspace:*",
     "dom-to-image": "^2.6.0",
     "eventsource-parser": "^0.1.0",
     "nanoid": "^4.0.1",
@@ -27,8 +28,9 @@
     "@unocss/preset-icons": "^0.50.6",
     "@unocss/preset-wind": "^0.50.6",
     "@vitejs/plugin-vue": "^4.0.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.0.4",
     "vite": "^4.1.0",
+    "vite-plugin-pwa": "^0.14.7",
     "vue-tsc": "^1.0.24"
   }
 }
