--- apps/mobile/src/utils/inputPanelHeight.ts
@@ -1,3 +1,0 @@
-import { ref } from "vue";
-
-export const inputPanelHeight = ref(0);
