+++ apps/mobile/vite.config.ts
@@ -4,6 +4,7 @@ import UnoCSS from "unocss/vite";
 import presetIcons from "@unocss/preset-icons";
 import presetWind from "@unocss/preset-wind";
 import presetAttributify from "@unocss/preset-attributify";
+import { VitePWA } from "vite-plugin-pwa";
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
@@ -18,6 +19,30 @@ export default defineConfig({
         }),
         presetWind(),
       ],
+    }),
+    VitePWA({
+      registerType: "autoUpdate",
+      devOptions: {
+        enabled: true,
+      },
+      manifest: {
+        name: "IM Chat",
+        short_name: "IM Chat",
+        description: "A chat tool",
+        theme_color: "#ffffff",
+        icons: [
+          {
+            src: "icon_192.png",
+            sizes: "192x192",
+            type: "image/png",
+          },
+          {
+            src: "icon_512.png",
+            sizes: "512x512",
+            type: "image/png",
+          },
+        ],
+      },
     }),
   ],
 });
