+++ apps/pc/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "chat-logic": "workspace:*",
     "@vueuse/core": "^9.13.0",
     "axios": "^1.3.4",
     "dom-to-image": "^2.6.0",
@@ -30,7 +31,7 @@
     "@unocss/preset-icons": "^0.50.6",
     "@unocss/preset-wind": "^0.50.6",
     "@vitejs/plugin-vue": "^4.0.0",
-    "typescript": "^4.9.5",
+    "typescript": "^5.0.4",
     "vite": "^4.1.0",
     "vue-tsc": "^1.0.24"
   }
