+++ apps/pc/src/components/message/MessageView.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { h, onMounted, ref, watch } from "vue";
 import { useElementHover } from "@vueuse/core";
-import { injectTreeChat, MessageNode } from "../../logic/TreeChat";
+import { injectTreeChat, MessageNode } from "chat-logic";
 import { useDialog } from "naive-ui";
 import ShareLink from "../share/ShareLink.vue";
 import InfoBox from "./InfoBox.vue";
