+++ packages/chat-logic/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "chat-logic",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.ts",
+  "types": "index.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "@vueuse/core": "^9.13.0",
+    "eventsource-parser": "^0.1.0",
+    "nanoid": "^4.0.1",
+    "typescript": "^5.0.4",
+    "vue": "^3.2.45"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
