+++ packages/chat-logic/src/utils/color.ts
