+++ packages/chat-logic/src/utils/sse.ts
